Moved JIT config specific code from api_test.cpp to api_test_jit.cpp
